ml_service/requirements.txt
# Core FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
pymongo==4.6.0
motor==3.3.2

# Data processing and analysis
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2

# Natural Language Processing
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1

# Machine Learning
tensorflow==2.15.0
torch==2.1.1
transformers==4.36.2

# Data visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production
gunicorn==21.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4